+++ R/select.R
@@ -179,6 +179,9 @@ summary.ctmm.single <- function(object, level=0.95, level.UD=0.95, units=TRUE, .
     ms <- ms + MSPEED$EST
     var.ms <- var.ms + MSPEED$VAR
 
+    # chi^2 degrees of freedom
+    DOF.speed <- 2*ms^2/var.ms
+
     # root mean square velocity
     # pretty units
     rms <- sqrt(ms)
@@ -192,6 +195,8 @@ summary.ctmm.single <- function(object, level=0.95, level.UD=0.95, units=TRUE, .
     par <- rbind(par,rms)
     rownames(par)[nrow(par)] <- "speed"
   }
+  else
+  { DOF.speed <- 0 }
 
   # did we estimate errors?
   if("error" %in% rownames(object$COV))
@@ -232,8 +237,8 @@ summary.ctmm.single <- function(object, level=0.95, level.UD=0.95, units=TRUE, .
   # only valid for processes with a stationary mean
   if(object$range)
   {
-    SUM$DOF <- c( DOF.mean(object) , DOF.area(object) )
-    names(SUM$DOF) <- c( "mean","area")
+    SUM$DOF <- c( DOF.mean(object) , DOF.area(object) , DOF.speed )
+    names(SUM$DOF) <- c("mean","area","speed")
   }
 
   SUM$CI <- par
@@ -251,6 +256,14 @@ sort.ctmm <- function(x, decreasing=FALSE, IC="AICc", ...)
   ICS <- sapply(x,function(m){m[[IC]]})
   IND <- sort(ICS,method="quick",index.return=TRUE,decreasing=decreasing)$ix
   x <- x[IND]
+  return(x)
+}
+
+min.ctmm <- function(x,IC="AICc",...)
+{
+  ICS <- sapply(x,function(m){m[[IC]]})
+  MIN <- which.min(ICS)
+  return(x[[MIN]])
 }
 
 #########
@@ -328,6 +341,9 @@ alpha.ctmm <- function(CTMM,alpha)
 ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",trace=FALSE,...)
 {
   alpha <- 1-level
+  trace2 <- if(trace) { trace-1 } else { 0 }
+
+  UERE <- get.error(data,CTMM,flag=TRUE) # error flag only
 
   drift <- get(CTMM$mean)
   if(CTMM$mean=="periodic")
@@ -336,28 +352,112 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",trace=FALSE
     message("Nyquist frequency estimated at harmonic ",paste(Nyquist,collapse=" ")," of the period.")
   }
 
-  trace2 <- if(trace) { trace-1 } else { 0 }
-  # fit the intital guess
-  if(trace) { message("* Fitting models ",name.ctmm(CTMM)) }
-
-  CTMM <- ctmm.fit(data,CTMM,trace=trace2,...)
-  OLD <- ctmm()
-  MODELS <- list(CTMM)
-  # while progress is being made, keep going, keep going, ...
-  while(!identical(CTMM,OLD))
+  # initial guess in case of pREML (better for optimization)
+  get.mle <- function(FIT=CTMM)
   {
-    GUESS <- list()
-    beta <- alpha.ctmm(CTMM,alpha)
-
-    # initial guess in case of pREML (better for optimization)
-    MLE <- CTMM
+    MLE <- FIT
     if(!get("EMPTY",pos=MLE.env)) # will have been set from ctmm.fit first run
     {
       MLE <- get("MLE",pos=MLE.env)
       # check that structure is consistent
-      if(is.null(MLE) || name.ctmm(MLE)!=name.ctmm(CTMM)) { MLE <- CTMM }
+      if(is.null(MLE) || name.ctmm(MLE)!=name.ctmm(FIT)) { MLE <- FIT }
+    }
+    return(MLE)
+  }
+
+  # function to simplify complexity of models
+  simplify <- function(M,par)
+  {
+    if("eccentricity" %in% par)
+    {
+      M$isotropic <- TRUE
+      M$sigma <- covm(M$sigma,isotropic=TRUE)
+    }
+    if("circle" %in% par) { M$circle <- FALSE }
+
+    return(M)
+  }
+
+  # consider a bunch of new models and update best model without duplication
+  iterate <- function(GUESS)
+  {
+    # name the proposed models
+    names(GUESS) <- sapply(GUESS,name.ctmm)
+    # remove models alreadt fit
+    RM <- which(names(GUESS) %in% names(MODELS))
+    if(length(RM)) { GUESS <- GUESS[-RM] }
+
+    # fit every model
+    if(trace && length(GUESS)) { message("* Fitting models ",paste(names(GUESS),collapse=", ")) }
+    #? should I run select here instead of fit ?
+    GUESS <- lapply(GUESS,function(g){ctmm.fit(data,g,trace=trace2,...)})
+    MODELS <<- c(MODELS,GUESS)
+
+    # what is the new best model?
+    OLD <<- CTMM
+    CTMM <<- min.ctmm(MODELS)
+  }
+
+  ########################
+  # PHASE 1: work our way up to complicated autocorrelation models
+  # all of the features we need to fit numerically
+  FEATURES <- id.parameters(CTMM,UERE=UERE)$NAMES
+  # consider only features unnecessary "compatibility"
+  FEATURES <- FEATURES[!(FEATURES=="area")]
+  FEATURES <- FEATURES[!(FEATURES=="error")]
+  FEATURES <- FEATURES[!grepl("tau",FEATURES)]
+
+  # start with the most basic "compatible" model
+  GUESS <- simplify(CTMM,FEATURES)
+  if(trace) { message("* Fitting model ",name.ctmm(GUESS)) }
+  TARGET <- CTMM
+  CTMM <- ctmm.fit(data,GUESS,trace=trace2,...)
+  MODELS <- list(CTMM)
+  names(MODELS) <- sapply(MODELS,name.ctmm)
+
+  OLD <- ctmm()
+  while(!identical(CTMM,OLD))
+  {
+    GUESS <- list()
+    MLE <- get.mle()
+
+    # consider non-zero eccentricity
+    if(("eccentricity" %in% FEATURES) && MLE$isotropic)
+    {
+      GUESS <- c(GUESS,list(MLE))
+      n <- length(GUESS)
+      GUESS[[n]]$isotropic <- FALSE
+      # copy over target angle, but leave eccentricity zero to start (featureless)
+      sigma <- attr(GUESS[[n]]$sigma,"par")
+      sigma["angle"] <- attr(TARGET$sigma,"par")['angle']
+      sigma <- covm(sigma,isotropic=FALSE,axes=TARGET$axes)
+      GUESS[[n]]$sigma <- sigma
     }
 
+    # consider circulation
+    if(("circle" %in% FEATURES) && !MLE$circle)
+    {
+      GUESS <- c(GUESS,list(MLE))
+      GUESS[[length(GUESS)]]$circle <- 2 * .Machine$double.eps * sign(TARGET$circle)
+    }
+
+    # consider no error?
+    #
+
+    # consider a bunch of new models and update best model without duplication
+    iterate(GUESS)
+  }
+
+  #############################
+  # PHASE 2: work our way down to simpler autocorrelation models & work our way up to more complex trend models
+  OLD <- ctmm()
+  # CTMM <- min.ctmm(MODELS)
+  while(!identical(CTMM,OLD))
+  {
+    GUESS <- list()
+    MLE <- get.mle()
+
+    beta <- alpha.ctmm(CTMM,alpha)
     # consider if some timescales are actually zero
     CI <- confint.ctmm(CTMM,alpha=beta)
 
@@ -366,8 +466,8 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",trace=FALSE
       Q <- CI["tau velocity",1]
       if(is.nan(Q) || (Q<=0))
       {
-         GUESS[[length(GUESS)+1]] <- MLE
-         GUESS[[length(GUESS)]]$tau <- MLE$tau[-length(MLE$tau)]
+        GUESS <- c(GUESS,list(MLE))
+        GUESS[[length(GUESS)]]$tau <- MLE$tau[-length(MLE$tau)]
       }
     }
     else if(length(CTMM$tau)==1)
@@ -375,31 +475,16 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",trace=FALSE
       Q <- CI["tau position",1]
       if(is.nan(Q) || (Q<=0))
       {
-        GUESS[[length(GUESS)+1]] <- MLE
+        GUESS <- c(GUESS,list(MLE))
         GUESS[[length(GUESS)]]$tau <- NULL
       }
     }
-
-    # consider if there is no telemetry error
-    # if(CTMM$error)
-    # {
-    #   Q <- CI["error",1]
-    #   if(is.nan(Q) || (Q<=0)) # This will never happen wich chi-square error model
-    #   {
-    #     GUESS[[length(GUESS)+1]] <- MLE
-    #     GUESS[[length(GUESS)]]$error <- FALSE
-    #   }
-    # }
 
     # consider if there is no circulation
     if(CTMM$circle)
     {
       Q <- CI["circle",3]
-      if(is.nan(Q) || (Q==Inf))
-      {
-        GUESS[[length(GUESS)+1]] <- MLE
-        GUESS[[length(GUESS)]]$circle <- FALSE
-      }
+      if(is.nan(Q) || (Q==Inf)) { GUESS <- c(GUESS,list(simplify(MLE,"circle"))) }
     }
 
     # consider if eccentricity is zero
@@ -407,34 +492,24 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",trace=FALSE
     {
       Q <- "eccentricity"
       Q <- stats::qnorm(beta/2,mean=CTMM$sigma@par[Q],sd=sqrt(CTMM$COV[Q,Q]))
-      if(Q <= 0)
-      {
-        GUESS[[length(GUESS)+1]] <- MLE
-        GUESS[[length(GUESS)]]$isotropic <- TRUE
-        GUESS[[length(GUESS)]]$sigma <- covm(MLE$sigma,isotropic=T)
-      }
+      if(Q <= 0) { GUESS <- c(GUESS,list(simplify(MLE,"eccentricity"))) }
     }
 
     # consider if the mean could be more detailed
     GUESS <- c(GUESS,drift@refine(MLE))
 
-    # fit every model
-    if(trace && length(GUESS)) { message("* Fitting models ",paste(sapply(GUESS,name.ctmm),collapse=", ")) }
-    GUESS <- lapply(GUESS,function(g){ctmm.fit(data,g,trace=trace2,...)})
-    MODELS <- c(MODELS,GUESS)
-
-    # what is the new best model?
-    OLD <- CTMM
-    MODELS <- sort.ctmm(MODELS)
-    CTMM <- MODELS[[1]]
+    # consider a bunch of new models and update best model without duplication
+    iterate(GUESS)
   }
 
-  # name all of the models
-  names(MODELS) <- sapply(MODELS,name.ctmm)
-
   # return the best or return the full list of models
-  if(verbose) { return(MODELS) }
-  else { return(MODELS[[1]]) }
+  if(verbose)
+  {
+    MODELS <- sort.ctmm(MODELS)
+    return(MODELS)
+  }
+  else
+  { return(CTMM) }
 }